LayoutTests/imported/w3c/web-platform-tests/wpt_internal/html/canvas/drawElementImage/privacy/filtered-div-ref-svg-feimage-svg-ignored.https.sub-expected.txt



FAIL drawElementImage does not use cross-origin feimage using SVG-as-Image via CSS assert_array_equals: Cross origin should not draw expected property 0 to be 0 but got 255 (expected array [0, 128, 0, 255] got object "255,0,0,255")
